discover-smarthome-backend/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.9
cloudinary==1.28.0
cssselect==1.1.0
cssutils==2.3.0
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
lxml==4.7.1
Pillow==8.4.0
premailer==3.10.0
psycopg2==2.9.2
PyJWT==2.3.0
pytz==2021.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.28
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
whitenoise==5.3.0
yagmail==0.14.260
